requirements.txt
# Core dependencies
# torch>=1.12.0
# torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
Pillow>=8.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Data processing
pandas>=1.3.0
opencv-python>=4.5.0
tqdm>=4.62.0

# Configuration and logging  
PyYAML>=5.4.0
tensorboard>=2.7.0
wandb>=0.12.0

# Graph neural networks
torch-geometric>=2.0.0
networkx>=2.6.0

# LoRA and parameter-efficient fine-tuning
peft>=0.4.0
transformers>=4.20.0

# Evaluation and metrics
faiss-cpu>=1.7.0  # For efficient similarity search
timm>=0.6.0  # For pretrained models

# Development tools
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
plotly>=5.0.0  # For interactive plots

# XML parsing (for VeRi dataset labels)
lxml>=4.6.0
xmltodict>=0.12.0

# Memory monitoring
psutil>=5.8.0
gpustat>=1.0.0